requirements.txt
requests 
pdf2image 
pillow
python-dotenv
openai
fastapi
uvicorn
python-multipart
